Modify Note App Data Structure
Started creating data structure. It is quite
difficult and having problems thinking of the best
way to store the notes across all views. So far,
app is able to push up to 2 compose views and
be able to show them in the main view controller.
I need to figure out a better way to save the
notes across all views.