Update Note Storage Structure
Finally, after a long time figuring out a way to
save notes properly, I think I've done it. Every
time a new note is loaded, it adds itself to
allNotes (in viewDidLoad), and the delegate
(main view) reassigns the value of allNotes after
it is changed. Also made the note controller have
optional note value so that when editing, you
can have note data that was created before.