Adjust for Keyboard and Trailing Actions
Override trailingSwipeActionConfigurationForRowAt
to implement actions a user can take when swiping
right on a cell. So far, they can share (activity
controller) and delete the individual cell.
Additionally, copied and pasted keyboard
function of making sure the text is in frame from
previous project. I still don't really understand
it... But, app is functional for now!!