Open Random Note, Compose, Multiple Selection
Implement each of these features. Opening Random
note and compose were relatively easy, but I had
to look up documentation on multiple selection as
well as stack overflow. Apparently, you have to
put the table in edit mode, and for the cell class
you have to set the accessory type to .checkmark
or .none depending on whether the cell is selected
during the table's edit mode.