Selection for Collection View & Deletion
Selection was difficult for collection view.
Unlike table view, there is no setEdit that
broadcast to all cells that the view is in edit
mode, so every time the user presses 'select note'
it flags a editMode variable in each collection
view cell that it is in edit mode. Also, since
there is no accessory member (unlike table cells),
I had to cerate a custom overlay to the view with
an empty circle signifying it is not selected,
and a checkmark and colored border to signify it
has. ALL OF THIS WAS SO HARD FOR WHAT BRUH WHY
CANT APPLE JUST MAKE IT LIKE THEIR TABLE VIEW.
Anyways, also updated deletion for the collection
cells.
Next goals:
1. actually having sections for collection view
2. snapshot of view for the cell
3. look at more features to implement
4. make UI snappier
5. ;aldskfa;sldjfa;sdjfads;lkfja (lol)