Section Labels for Collection View
Adding section labels for a collection view is
significantly more difficult than table view.
You need to register a UICollectionReusableView
that is associated with supplementary view of some
kind as well as a reuse identifier. When a new
section is created, it collection view
viewForSupplementaryElementOfKind to make a view
that is presented as a header. Anyways, my brain
is fried for today. Maybe I'll take a short rest.