Modify setup functions for views, Passing Notes
A note is passed to collection view cells now and
modified the setup functions to be more modular
instead of all being in one place. I tried adding
snapshots to collection view cells, but it was too
difficult and I think I'll try later. For now, I
think this is a good start to a clone of the note
taking app.