Project 23 Setup & Enemy Creation
Set up game scene. The game is akin to fruit ninja
so the goal is to slice penguins, but not bombs.
The enemy creation was by far the most difficult
part so far. Needed to create random x and y
velocities, as well as enemy positions, and enemy
types. However, needed to create container for
the bombs because of the emitter nodes for their
fuses. Used AVAudioPlayer to play the bomb sound
effect because the SKAction way of playing sound
doesn't allow you to stop the sound abruptly.