Score Keeping (End Game), Auto Layout, Adding Wind
The game now ends after a player has scored 3
times with a giant UILabel saying "Game Over".
Added auto layout rules that work decently with
iPhones of different sizes, but not iPads...
Finally, added wind by changing physicsWorld.
gravity.dx whenever switchPlayer is called. The
corresponding UILabel is changed to notify the
player which was the wind is blowing.