Debugging/Tracking Performance Issues
Use Instruments to track performance when the user
opens the app and as they use it. As a result,
instead of redrawing shadows for the table cell
image, implemented a 1 pass system when calling
renderer.image to both scale the image and create
shadows which saved on performance. Additionally,
tracked memory allocation of the app and had to
reconfigure how cells are dequeued (was previously
just creating new cells) and had to register cells
programatically in viewDidLoad.