Handle Force Unwraps, Allocation Instrument, Cache
Removed all force unwraps, mostly with a bunch of
guard let statements. Used the allocation
instrument on my previous project, Simple Notes,
to check if there was an persistent memory.
Finally, created a cache of images upon the app
loading so when cellForRowAt is called, it uses
those images instead of redrawing it every time.